# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZKTeco Relay API
  version: 1.2.0
  description: |
    Windows ZKTeco 脱机通讯 REST/WebSocket 中继服务。
    除健康检查和文档端点外，所有请求均需提供 X-API-Key。
servers:
  - url: http://127.0.0.1:5080
    description: 默认本机服务
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/devices:
    get:
      tags: [Devices]
      operationId: listDevices
      summary: 获取设备会话列表
      responses:
        '200':
          description: 设备状态列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Devices]
      operationId: getDevice
      summary: 获取单个设备状态
      responses:
        '200':
          description: 设备状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
  /api/v1/devices/{deviceId}/connect:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      operationId: connectDevice
      summary: 连接设备
      description: 连接成功或失败都会保存连接配置；成功后启用自动连接。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectDeviceRequest'
            example:
              ipAddress: 192.168.1.100
              port: 4370
              communicationPassword: ''
      responses:
        '200':
          description: 连接成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConnectionResult'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
        '502':
          $ref: '#/components/responses/DeviceOperationFailed'
  /api/v1/devices/{deviceId}/disconnect:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      operationId: disconnectDevice
      summary: 断开设备并关闭自动连接
      responses:
        '200':
          description: 已断开
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisconnectResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
  /api/v1/devices/{deviceId}/restart:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [Devices]
      operationId: restartDevice
      summary: 重启设备
      responses:
        '202':
          description: 重启命令已发送
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestartResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
  /api/v1/devices/{deviceId}/attendance:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Attendance]
      operationId: readAllAttendance
      summary: 拉取全部普通考勤记录
      description: 兼容端点，响应体保持为数组。
      responses:
        '200':
          description: 考勤记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/DeviceNotFound'
        '409':
          $ref: '#/components/responses/DeviceUnavailable'
...
  /api/v1/events/ws:
    get:
      tags: [Events]
      operationId: subscribeRealtimeEvents
      summary: 订阅实时事件 WebSocket
      description: |
        使用 WebSocket 升级连接。服务端客户端应使用 X-API-Key；
        浏览器可使用 apiKey 查询参数。可按 deviceId 和 eventType 过滤。
        afterSequence 用于从 SQLite 按序补发后继续实时推送。
      x-websocket: true
      x-websocket-message-schema:
        $ref: '#/components/schemas/RealtimeEvent'
      security:
        - ApiKeyAuth: []
        - ApiKeyQuery: []
      parameters:
        - name: apiKey
          in: query
          description: 浏览器 WebSocket 无法设置自定义请求头时使用。
          schema:
            type: string
        - name: deviceId
          in: query
          schema:
            type: string
        - name: eventType
          in: query
          description: 逗号分隔的事件类型。
          schema:
            type: string
          example: attendance,door,alarm
        - name: afterSequence
          in: query
          description: 已提交的最后一个事件序号，使用十进制字符串。
          schema:
            type: string
            pattern: '^[0-9]+$'
          example: '1024'
      responses:
        '101':
          description: WebSocket 协议升级成功，消息模型为 RealtimeEvent。
        '400':
          description: 请求不是 WebSocket 升级请求
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      description: 调用方定义的稳定设备标识，最长 64 个字符。
      schema:
        type: string
        maxLength: 64
      example: office-gate-01
...
  responses:
    Unauthorized:
      description: 缺少或提供了错误的 API Key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    DeviceNotFound:
      description: 设备会话不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: device_not_found
            message: Device was not found.
            vendorErrorCode: null
...
    DeviceUnavailable:
      description: 设备会话存在但尚未连接
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    DeviceOperationFailed:
      description: 厂商 SDK 操作失败或返回无效数据
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        vendorErrorCode:
          type: integer
          nullable: true
...
    ConnectDeviceRequest:
      type: object
      required: [ipAddress]
      properties:
        ipAddress:
          type: string
          description: 设备 IPv4 或 IPv6 地址。
        port:
          type: integer
          minimum: 1
          maximum: 65535
          default: 4370
        communicationPassword:
          type: string
          nullable: true
          description: 设备通信密码，不是 Relay API Key。
...
    DeviceConnectionResult:
      type: object
      required: [deviceId, connected]
      properties:
        deviceId:
          type: string
        connected:
          type: boolean
        error:
          type: string
          nullable: true
        vendorErrorCode:
          type: integer
          nullable: true
    DeviceStatus:
      type: object
      required: [deviceId, ipAddress, port, connected]
      properties:
        deviceId:
          type: string
        ipAddress:
          type: string
        port:
          type: integer
        connected:
          type: boolean
        connectedAt:
          type: string
          format: date-time
          nullable: true
        lastError:
          type: string
          nullable: true
        lastCommunicationAt:
          type: string
          format: date-time
          nullable: true
        disconnectedAt:
          type: string
          format: date-time
          nullable: true
        reconnectAttempt:
          type: integer
          nullable: true
        nextReconnectAt:
          type: string
          format: date-time
          nullable: true
...
    DisconnectResult:
      type: object
      required: [deviceId, connected]
      properties:
        deviceId:
          type: string
        connected:
          type: boolean
          enum: [false]
    RestartResult:
      type: object
      required: [deviceId, restarted]
      properties:
        deviceId:
          type: string
        restarted:
          type: boolean
    AttendanceRecord:
      type: object
      required: [enrollNumber, verifyMode, inOutMode, timestamp, workCode]
      properties:
        enrollNumber:
          type: string
        verifyMode:
          type: integer
          description: 厂商 SDK 原始验证方式。
        inOutMode:
          type: integer
          description: 厂商 SDK 原始考勤状态。
        timestamp:
          type: string
          format: date-time
        workCode:
          type: integer
...
    RealtimeEvent:
      type: object
      required: [eventId, deviceId, eventType, occurredAt, data]
      properties:
        eventId:
          type: string
        eventSequence:
          type: string
          nullable: true
          pattern: '^[1-9][0-9]*$'
          description: 持久事件的 SQLite 单调递增序号；控制消息可为空。
        deviceId:
          type: string
        eventType:
          type: string
        occurredAt:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true